third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-color-mix-function-expected.txt

This is a testharness.js-based test.
Found 172 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property color value 'color-mix(in lch, white, blue)'
  Colors do not match.\nActual:   lch(64.7821 65.6161 328.997)\nExpected: lch(64.78 65.6 301.37).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 301.37 +/- 0.1, expected 301.37 but got 328.997
[FAIL] Property color value 'color-mix(in lch, white 10%, blue)'
  Colors do not match.\nActual:   lch(36.6087 118.094 306.894)\nExpected: lch(36.61 118.09 301.37).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 301.37 +/- 0.1, expected 301.37 but got 306.894
[FAIL] Property color value 'color-mix(in oklch, white, blue)'
  Colors do not match.\nActual:   oklch(0.725987 0.15663 323.92)\nExpected: oklch(0.726 0.156 264.052).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 264.052 +/- 0.01, expected 264.052 but got 323.92
[FAIL] Property color value 'color-mix(in oklch, white 10%, blue)'
  Colors do not match.\nActual:   oklch(0.506782 0.281894 276.025)\nExpected: oklch(0.506 0.281 264.052).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 264.052 +/- 0.01, expected 264.052 but got 276.025
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 25%, color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 25%, color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, 25% color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, 25% color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7) 25%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3), 25% color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3), 25% color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 25%, color(display-p3 .5 .6 .7) 75%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 25%, color(display-p3 .5 .6 .7) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 30%, color(display-p3 .5 .6 .7) 90%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 30%, color(display-p3 .5 .6 .7) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 12.5%, color(display-p3 .5 .6 .7) 37.5%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 12.5%, color(display-p3 .5 .6 .7) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 0%, color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3) 0%, color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .5), color(display-p3 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .5), color(display-p3 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 25%, color(display-p3 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 25%, color(display-p3 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, 25% color(display-p3 .1 .2 .3 / .4), color(display-p3 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in display-p3, 25% color(display-p3 .1 .2 .3 / .4), color(display-p3 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4), color(display-p3 .5 .6 .7 / .8) 25%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4), color(display-p3 .5 .6 .7 / .8) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4), 25% color(display-p3 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4), 25% color(display-p3 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 25%, color(display-p3 .5 .6 .7 / .8) 75%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 25%, color(display-p3 .5 .6 .7 / .8) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 30%, color(display-p3 .5 .6 .7 / .8) 90%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 30%, color(display-p3 .5 .6 .7 / .8) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 12.5%, color(display-p3 .5 .6 .7 / .8) 37.5%)'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 12.5%, color(display-p3 .5 .6 .7 / .8) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 0%, color(display-p3 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / .4) 0%, color(display-p3 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, transparent, color(display-p3 0.3 0.4 0.5))'
  assert_true: 'color-mix(in display-p3, transparent, color(display-p3 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, transparent 10%, color(display-p3 0.3 0.4 0.5))'
  assert_true: 'color-mix(in display-p3, transparent 10%, color(display-p3 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 0.1 0.2 0.3 / 0), color(display-p3 0.3 0.4 0.5))'
  assert_true: 'color-mix(in display-p3, color(display-p3 0.1 0.2 0.3 / 0), color(display-p3 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 0.1 0.2 0.3 / 0) 10%, color(display-p3 0.3 0.4 0.5))'
  assert_true: 'color-mix(in display-p3, color(display-p3 0.1 0.2 0.3 / 0) 10%, color(display-p3 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 1 1 1), color(display-p3 0 0 1))'
  assert_true: 'color-mix(in display-p3, color(display-p3 1 1 1), color(display-p3 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 1 1 1) 10%, color(display-p3 0 0 1))'
  assert_true: 'color-mix(in display-p3, color(display-p3 1 1 1) 10%, color(display-p3 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 2 3 4 / 5), color(display-p3 4 6 8 / 10))'
  assert_true: 'color-mix(in display-p3, color(display-p3 2 3 4 / 5), color(display-p3 4 6 8 / 10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 -2 -3 -4), color(display-p3 -4 -6 -8))'
  assert_true: 'color-mix(in display-p3, color(display-p3 -2 -3 -4), color(display-p3 -4 -6 -8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 -2 -3 -4 / -5), color(display-p3 -4 -6 -8 / -10))'
  assert_true: 'color-mix(in display-p3, color(display-p3 -2 -3 -4 / -5), color(display-p3 -4 -6 -8 / -10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 none none none), color(display-p3 none none none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 none none none), color(display-p3 none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 none none none), color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 none none none), color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 none none none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 none), color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 none), color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 .5 .6 none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3), color(display-p3 .5 .6 none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 none .2 .3), color(display-p3 .5 none .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 none .2 .3), color(display-p3 .5 none .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / none), color(display-p3 .5 .6 .7))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / none), color(display-p3 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / none), color(display-p3 .5 .6 .7 / 0.5))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / none), color(display-p3 .5 .6 .7 / 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / none), color(display-p3 .5 .6 .7 / none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / none), color(display-p3 .5 .6 .7 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 0.5 none none / none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 0.5 none none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 none 0.5 none / none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 none 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 none none 0.5 / none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 none none 0.5 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 0.5 0.5 none / none))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 0.5 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 none none none / 50%))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 none none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 0.5 none none / 50%))'
  assert_true: 'color-mix(in display-p3, color(display-p3 .1 .2 .3 / 25%) 0%, color(display-p3 0.5 none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 25%, color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 25%, color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, 25% color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, 25% color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7) 25%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), 25% color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), 25% color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 25%, color(a98-rgb .5 .6 .7) 75%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 25%, color(a98-rgb .5 .6 .7) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 30%, color(a98-rgb .5 .6 .7) 90%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 30%, color(a98-rgb .5 .6 .7) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 12.5%, color(a98-rgb .5 .6 .7) 37.5%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 12.5%, color(a98-rgb .5 .6 .7) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 0%, color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3) 0%, color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .5), color(a98-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .5), color(a98-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 25%, color(a98-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 25%, color(a98-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, 25% color(a98-rgb .1 .2 .3 / .4), color(a98-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in a98-rgb, 25% color(a98-rgb .1 .2 .3 / .4), color(a98-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4), color(a98-rgb .5 .6 .7 / .8) 25%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4), color(a98-rgb .5 .6 .7 / .8) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4), 25% color(a98-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4), 25% color(a98-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 25%, color(a98-rgb .5 .6 .7 / .8) 75%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 25%, color(a98-rgb .5 .6 .7 / .8) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 30%, color(a98-rgb .5 .6 .7 / .8) 90%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 30%, color(a98-rgb .5 .6 .7 / .8) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 12.5%, color(a98-rgb .5 .6 .7 / .8) 37.5%)'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 12.5%, color(a98-rgb .5 .6 .7 / .8) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 0%, color(a98-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / .4) 0%, color(a98-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, transparent, color(a98-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in a98-rgb, transparent, color(a98-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, transparent 10%, color(a98-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in a98-rgb, transparent 10%, color(a98-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb 0.1 0.2 0.3 / 0), color(a98-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb 0.1 0.2 0.3 / 0), color(a98-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb 0.1 0.2 0.3 / 0) 10%, color(a98-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb 0.1 0.2 0.3 / 0) 10%, color(a98-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb 1 1 1), color(a98-rgb 0 0 1))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb 1 1 1), color(a98-rgb 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb 1 1 1) 10%, color(a98-rgb 0 0 1))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb 1 1 1) 10%, color(a98-rgb 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb 2 3 4 / 5), color(a98-rgb 4 6 8 / 10))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb 2 3 4 / 5), color(a98-rgb 4 6 8 / 10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb -2 -3 -4), color(a98-rgb -4 -6 -8))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb -2 -3 -4), color(a98-rgb -4 -6 -8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb -2 -3 -4 / -5), color(a98-rgb -4 -6 -8 / -10))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb -2 -3 -4 / -5), color(a98-rgb -4 -6 -8 / -10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb none none none), color(a98-rgb none none none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb none none none), color(a98-rgb none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb none none none), color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb none none none), color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb none none none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 none), color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 none), color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb none .2 .3), color(a98-rgb .5 none .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb none .2 .3), color(a98-rgb .5 none .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / none), color(a98-rgb .5 .6 .7))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / none), color(a98-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / none), color(a98-rgb .5 .6 .7 / 0.5))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / none), color(a98-rgb .5 .6 .7 / 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / none), color(a98-rgb .5 .6 .7 / none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / none), color(a98-rgb .5 .6 .7 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb 0.5 none none / none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb 0.5 none none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb none 0.5 none / none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb none 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb none none 0.5 / none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb none none 0.5 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb 0.5 0.5 none / none))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb 0.5 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb none none none / 50%))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb none none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb 0.5 none none / 50%))'
  assert_true: 'color-mix(in a98-rgb, color(a98-rgb .1 .2 .3 / 25%) 0%, color(a98-rgb 0.5 none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 25%, color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 25%, color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, 25% color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, 25% color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7) 25%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), 25% color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), 25% color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 25%, color(prophoto-rgb .5 .6 .7) 75%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 25%, color(prophoto-rgb .5 .6 .7) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 30%, color(prophoto-rgb .5 .6 .7) 90%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 30%, color(prophoto-rgb .5 .6 .7) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 12.5%, color(prophoto-rgb .5 .6 .7) 37.5%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 12.5%, color(prophoto-rgb .5 .6 .7) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 0%, color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3) 0%, color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .5), color(prophoto-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .5), color(prophoto-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 25%, color(prophoto-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 25%, color(prophoto-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, 25% color(prophoto-rgb .1 .2 .3 / .4), color(prophoto-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in prophoto-rgb, 25% color(prophoto-rgb .1 .2 .3 / .4), color(prophoto-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4), color(prophoto-rgb .5 .6 .7 / .8) 25%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4), color(prophoto-rgb .5 .6 .7 / .8) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4), 25% color(prophoto-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4), 25% color(prophoto-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 25%, color(prophoto-rgb .5 .6 .7 / .8) 75%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 25%, color(prophoto-rgb .5 .6 .7 / .8) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 30%, color(prophoto-rgb .5 .6 .7 / .8) 90%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 30%, color(prophoto-rgb .5 .6 .7 / .8) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 12.5%, color(prophoto-rgb .5 .6 .7 / .8) 37.5%)'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 12.5%, color(prophoto-rgb .5 .6 .7 / .8) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 0%, color(prophoto-rgb .5 .6 .7 / .8))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / .4) 0%, color(prophoto-rgb .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, transparent, color(prophoto-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in prophoto-rgb, transparent, color(prophoto-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, transparent 10%, color(prophoto-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in prophoto-rgb, transparent 10%, color(prophoto-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb 0.1 0.2 0.3 / 0), color(prophoto-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb 0.1 0.2 0.3 / 0), color(prophoto-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb 0.1 0.2 0.3 / 0) 10%, color(prophoto-rgb 0.3 0.4 0.5))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb 0.1 0.2 0.3 / 0) 10%, color(prophoto-rgb 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb 1 1 1), color(prophoto-rgb 0 0 1))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb 1 1 1), color(prophoto-rgb 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb 1 1 1) 10%, color(prophoto-rgb 0 0 1))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb 1 1 1) 10%, color(prophoto-rgb 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb 2 3 4 / 5), color(prophoto-rgb 4 6 8 / 10))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb 2 3 4 / 5), color(prophoto-rgb 4 6 8 / 10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb -2 -3 -4), color(prophoto-rgb -4 -6 -8))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb -2 -3 -4), color(prophoto-rgb -4 -6 -8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb -2 -3 -4 / -5), color(prophoto-rgb -4 -6 -8 / -10))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb -2 -3 -4 / -5), color(prophoto-rgb -4 -6 -8 / -10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb none none none), color(prophoto-rgb none none none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb none none none), color(prophoto-rgb none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb none none none), color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb none none none), color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb none none none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 none), color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 none), color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb none .2 .3), color(prophoto-rgb .5 none .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb none .2 .3), color(prophoto-rgb .5 none .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / none), color(prophoto-rgb .5 .6 .7))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / none), color(prophoto-rgb .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / none), color(prophoto-rgb .5 .6 .7 / 0.5))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / none), color(prophoto-rgb .5 .6 .7 / 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / none), color(prophoto-rgb .5 .6 .7 / none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / none), color(prophoto-rgb .5 .6 .7 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb 0.5 none none / none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb 0.5 none none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb none 0.5 none / none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb none 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb none none 0.5 / none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb none none 0.5 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb 0.5 0.5 none / none))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb 0.5 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb none none none / 50%))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb none none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb 0.5 none none / 50%))'
  assert_true: 'color-mix(in prophoto-rgb, color(prophoto-rgb .1 .2 .3 / 25%) 0%, color(prophoto-rgb 0.5 none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 25%, color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 25%, color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, 25% color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, 25% color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7) 25%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3), 25% color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3), 25% color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 25%, color(rec2020 .5 .6 .7) 75%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 25%, color(rec2020 .5 .6 .7) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 30%, color(rec2020 .5 .6 .7) 90%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 30%, color(rec2020 .5 .6 .7) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 12.5%, color(rec2020 .5 .6 .7) 37.5%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 12.5%, color(rec2020 .5 .6 .7) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 0%, color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3) 0%, color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .5), color(rec2020 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .5), color(rec2020 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 25%, color(rec2020 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 25%, color(rec2020 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, 25% color(rec2020 .1 .2 .3 / .4), color(rec2020 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in rec2020, 25% color(rec2020 .1 .2 .3 / .4), color(rec2020 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4), color(rec2020 .5 .6 .7 / .8) 25%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4), color(rec2020 .5 .6 .7 / .8) 25%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4), 25% color(rec2020 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4), 25% color(rec2020 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 25%, color(rec2020 .5 .6 .7 / .8) 75%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 25%, color(rec2020 .5 .6 .7 / .8) 75%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 30%, color(rec2020 .5 .6 .7 / .8) 90%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 30%, color(rec2020 .5 .6 .7 / .8) 90%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 12.5%, color(rec2020 .5 .6 .7 / .8) 37.5%)'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 12.5%, color(rec2020 .5 .6 .7 / .8) 37.5%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 0%, color(rec2020 .5 .6 .7 / .8))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / .4) 0%, color(rec2020 .5 .6 .7 / .8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, transparent, color(rec2020 0.3 0.4 0.5))'
  assert_true: 'color-mix(in rec2020, transparent, color(rec2020 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, transparent 10%, color(rec2020 0.3 0.4 0.5))'
  assert_true: 'color-mix(in rec2020, transparent 10%, color(rec2020 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 0.1 0.2 0.3 / 0), color(rec2020 0.3 0.4 0.5))'
  assert_true: 'color-mix(in rec2020, color(rec2020 0.1 0.2 0.3 / 0), color(rec2020 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 0.1 0.2 0.3 / 0) 10%, color(rec2020 0.3 0.4 0.5))'
  assert_true: 'color-mix(in rec2020, color(rec2020 0.1 0.2 0.3 / 0) 10%, color(rec2020 0.3 0.4 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 1 1 1), color(rec2020 0 0 1))'
  assert_true: 'color-mix(in rec2020, color(rec2020 1 1 1), color(rec2020 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 1 1 1) 10%, color(rec2020 0 0 1))'
  assert_true: 'color-mix(in rec2020, color(rec2020 1 1 1) 10%, color(rec2020 0 0 1))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 2 3 4 / 5), color(rec2020 4 6 8 / 10))'
  assert_true: 'color-mix(in rec2020, color(rec2020 2 3 4 / 5), color(rec2020 4 6 8 / 10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 -2 -3 -4), color(rec2020 -4 -6 -8))'
  assert_true: 'color-mix(in rec2020, color(rec2020 -2 -3 -4), color(rec2020 -4 -6 -8))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 -2 -3 -4 / -5), color(rec2020 -4 -6 -8 / -10))'
  assert_true: 'color-mix(in rec2020, color(rec2020 -2 -3 -4 / -5), color(rec2020 -4 -6 -8 / -10))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 none none none), color(rec2020 none none none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 none none none), color(rec2020 none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 none none none), color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 none none none), color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 none none none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 none none none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 none), color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 none), color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 .5 .6 none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3), color(rec2020 .5 .6 none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 none .2 .3), color(rec2020 .5 none .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 none .2 .3), color(rec2020 .5 none .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / none), color(rec2020 .5 .6 .7))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / none), color(rec2020 .5 .6 .7))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / none), color(rec2020 .5 .6 .7 / 0.5))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / none), color(rec2020 .5 .6 .7 / 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / none), color(rec2020 .5 .6 .7 / none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / none), color(rec2020 .5 .6 .7 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 0.5 none none / none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 0.5 none none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 none 0.5 none / none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 none 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 none none 0.5 / none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 none none 0.5 / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 0.5 0.5 none / none))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 0.5 0.5 none / none))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 none none none / 50%))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 none none none / 50%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 0.5 none none / 50%))'
  assert_true: 'color-mix(in rec2020, color(rec2020 .1 .2 .3 / 25%) 0%, color(rec2020 0.5 none none / 50%))' is a supported value for color. expected true got false
Harness: the test ran to completion.
